refactor(transformation): AJDA-2727 extract SessionVariablesExporter
Move DECLARE-name parsing, value normalisation, and result.json writing
out of Transformation into a dedicated SessionVariablesExporter so the
session-variable feature is a self-contained collaborator with its own
state. Transformation now composes the exporter, feeds it queries via
captureFromQuery() inside the run loop, and exposes exportSessionVariables()
as a thin delegate so Component's call site is unchanged.
